+++ lib/tty/box/border.rb
@@ -1,11 +1,17 @@
 # frozen_string_literal: true
 
 module TTY
-  module Box
+  class Box
     # A class reponsible for retrieving border options
     #
     # @api private
     class Border
+      BOX_CHARS = {
+        ascii: %w[+ + + + + + + + - | +],
+        light: %w[┘ ┐ ┌ └ ┤ ┴ ┬ ├ ─ │ ┼],
+        thick: %w[╝ ╗ ╔ ╚ ╣ ╩ ╦ ╠ ═ ║ ╬]
+      }.freeze
+
       BORDER_VALUES = %i[
         corner_bottom_right
         corner_top_right
@@ -20,11 +26,14 @@ module TTY
         cross
       ].freeze
 
+      # Parse border configuration
+      #
+      # @api public
       def self.parse(border)
         case border
         when Hash
           new(**border)
-        when *TTY::Box::BOX_CHARS.keys
+        when *BOX_CHARS.keys
           new(type: border)
         else
           raise ArgumentError,
@@ -63,6 +72,163 @@ module TTY
         @bottom       = check_name(:bottom, bottom)
         @bottom_left  = check_name(:bottom_left, bottom_left)
         @bottom_right = check_name(:bottom_right, bottom_right)
+      end
+
+      # Top border size
+      #
+      # @return [Integer]
+      #
+      # @api public
+      def top_size
+        top? ? 1 : 0
+      end
+
+      # Left border size
+      #
+      # @return [Integer]
+      #
+      # @api public
+      def left_size
+        left? ? 1 : 0
+      end
+
+      # Right border size
+      #
+      # @return [Integer]
+      #
+      # @api public
+      def right_size
+        right? ? 1 : 0
+      end
+
+      # Bottom border size
+      #
+      # @return [Integer]
+      #
+      # @api public
+      def bottom_size
+        bottom? ? 1 : 0
+      end
+
+      # Top left corner
+      #
+      # @return [String]
+      #
+      # @api public
+      def top_left_corner
+        return "" unless top_left? && left?
+
+        send(:"#{top_left}_char")
+      end
+
+      # Top right corner
+      #
+      # @return [String]
+      #
+      # @api public
+      def top_right_corner
+        return "" unless top_right? && right?
+
+        send(:"#{top_right}_char")
+      end
+
+      # Bottom left corner
+      #
+      # @return [String]
+      #
+      # @api public
+      def bottom_left_corner
+        return "" unless bottom_left? && left?
+
+        send(:"#{bottom_left}_char")
+      end
+
+      # Bottom right corner
+      #
+      # @return [String]
+      #
+      # @api public
+      def bottom_right_corner
+        return "" unless bottom_right? && right?
+
+        send(:"#{bottom_right}_char")
+      end
+
+      # Bottom right corner character
+      #
+      # @api public
+      def corner_bottom_right_char
+        BOX_CHARS[type][0]
+      end
+
+      # Top right corner character
+      #
+      # @api public
+      def corner_top_right_char
+        BOX_CHARS[type][1]
+      end
+
+      # Top left corner character
+      #
+      # @api public
+      def corner_top_left_char
+        BOX_CHARS[type][2]
+      end
+
+      # Bottom left corner character
+      #
+      # @api public
+      def corner_bottom_left_char
+        BOX_CHARS[type][3]
+      end
+
+      # Left divider character
+      #
+      # @api public
+      def divider_left_char
+        BOX_CHARS[type][4]
+      end
+
+      # Up divider character
+      #
+      # @api public
+      def divider_up_char
+        BOX_CHARS[type][5]
+      end
+
+      # Down divider character
+      #
+      # @api public
+      def divider_down_char
+        BOX_CHARS[type][6]
+      end
+
+      # Right divider character
+      #
+      # @api public
+      def divider_right_char
+        BOX_CHARS[type][7]
+      end
+
+      # Horizontal line character
+      #
+      # @api public
+      def line_char
+        BOX_CHARS[type][8]
+      end
+
+      # Vertical line character
+      #
+      # @api public
+      def pipe_char
+        BOX_CHARS[type][9]
+      end
+
+      # Intersection character
+      #
+      # @api public
+      def cross_char
+        BOX_CHARS[type][10]
       end
 
       private